Move code to part_nonrepl_common.cpp file
