+++ cloud/blockstore/libs/storage/partition_nonrepl/part_nonrepl_actor_base_request.cpp
@@ -54,7 +54,7 @@ bool TDiskAgentBaseRequestActor::HandleError(
     bool timedOut)
 {
     if (FAILED(error.GetCode())) {
-        ProcessError(ctx, *PartConfig, error);
+        ProcessError(*TActorContext::ActorSystem(), *PartConfig, error);
 
         Done(
             ctx,
